fix mapper and dto

+++ src/main/java/com/nexigroup/pagopa/cruscotto/service/dto/AnagPartnerDTO.java
@@ -36,7 +36,7 @@ public class AnagPartnerDTO implements Serializable {
 
     private LocalDate analysisPeriodEndDate;
     
-    private Integer stationsCount;
+    private Long stationsCount;
 
 //    private Integer associatedInstitutes;
 //
